Normalize whitespace in rails-post-deploy
The rails-post-deploy script unfortunately uses a mixtures of spaces
and 0x09 to represent indentation, so it's not even clear what local
convention to follow.  It appears from the mixture that the intention
is that the tabs should represent 4 spaces, so this commit replaces
them, strips trailing whitespace and changes some non-standard
indentation at the end of a bash heredoc.