Invoke script/load-sample-data with bundle exec
This fixes the error:

    rake aborted!
    uninitialized constant Rake::DSL
    /var/www/alaveteli/alaveteli/Rakefile:6
    (See full trace by running task with --trace)

... although it's not clear to me why, since inside the
load-sample-data script the code is run via
"bundle exec script runner".

An alternative solution, suggested on StackOverflow, is to put:

   require 'rake/dsl_definition'

... before "require 'rake'" in the Rakefile - this also
fixes the problem but causes a couple of errors to do with
redefining constants, and it would be annoying to have them
output on every rake task.